src/main/java/Hero.java
public class Hero {

    private String mName;
    private int mAge;
    private String mAbility;
    private String mWeakness;


    public Hero(String name, int age, String ability, String weakness) {
        mName = name;
        mAge = age;
        mAbility = ability;
        mWeakness = weakness;
    }

    public String getName() {
        return mName;
    }

    public String getAbility() {
        return mAbility;
    }

    public int getAge() {
        return mAge;
    }

    public String getWeakness() {
        return mWeakness;
    }
}
